+++ nodes/__init__.py
@@ -1,5 +1,5 @@
 """HomeKit Hub Node Server nodes."""
-VERSION = "0.1.12"
+VERSION = "0.1.13"
 from .Controller import Controller as Controller  # noqa: E402,F401
 
 __all__ = ["Controller", "VERSION"]